added most popular TV show table, data not getting added correctly

+++ api_data.py
@@ -21,7 +21,7 @@ def prepare_most_popular_shows(most_popular_data: list[dict]) -> list[tuple]:
         show_values = list(show_data.values())
         show_values[1] = int(show_values[1])  # convert rank to int
         show_values[5] = int(show_values[5])  # convert year to int
-        show_values[7] = int(show_values[7])
+        show_values[7] = float(show_values[7])
         show_values[8] = int(show_values[8])
         show_values = tuple(show_values)
         data_for_database.append(show_values)
